[Add] Add obstacle
Obstacle facotry made.
Printing player and obstacle hass been added; however, now the scene is automatically move to the next scene which is the Scene 2. So, it has to be fixed to move to the next scene by certain condition such as reaching to the goal.

Things to be done next:
1) multiple obstacles for horizontal and vertical line.
since I already started this.
2) Make condition to move to the next stage.
3) functions to move
( Already stated working on this but didnt make it. has error to read the map? or more likely,  reading a  x&y - axis varibles).